internal/log/segment.go
package log

import (
	"fmt"
	"os"
	"path/filepath"

	log_v1 "github.com/sazid/proglog/api/v1"
	"google.golang.org/protobuf/proto"
)

type segment struct {
	store                  *store
	index                  *index
	baseOffset, nextOffset uint64
	config                 Config
}

func newSegment(dir string, baseOffset uint64, c Config) (*segment, error) {
	s := &segment{
		baseOffset: baseOffset,
		config:     c,
	}

	storeFile, err := os.OpenFile(
		filepath.Join(dir, fmt.Sprintf("%d%s", baseOffset, ".store")),
		os.O_RDWR|os.O_CREATE|os.O_APPEND,
		0644,
	)
	if err != nil {
		return nil, fmt.Errorf("error creating/opening store file: %w", err)
	}
	s.store, err = newStore(storeFile)
	if err != nil {
		return nil, fmt.Errorf("error creating new store: %w", err)
	}

	indexFile, err := os.OpenFile(
		filepath.Join(dir, fmt.Sprintf("%d%s", baseOffset, ".index")),
		os.O_RDWR|os.O_CREATE,
		0644,
	)
	if err != nil {
		return nil, fmt.Errorf("error creating/opening index file: %w", err)
	}
	s.index, err = newIndex(indexFile, c)
	if err != nil {
		return nil, fmt.Errorf("error creating new index: %w", err)
	}

	if off, _, err := s.index.Read(-1); err != nil {
		s.nextOffset = baseOffset
	} else {
		s.nextOffset = baseOffset + uint64(off) + 1
	}

	return s, nil
}

func (s *segment) Append(record *log_v1.Record) (offset uint64, err error) {
	cur := s.nextOffset
	record.Offset = cur
	p, err := proto.Marshal(record)
	if err != nil {
		return 0, err
	}
	_, pos, err := s.store.Append(p)
	if err != nil {
		return 0, fmt.Errorf("error appending to store: %w", err)
	}
	if err := s.index.Write(
		// index offsets are relative to base offsets
		uint32(cur-s.baseOffset),
		pos,
	); err != nil {
		return 0, fmt.Errorf("error writing to index: %w", err)
	}
	s.nextOffset++
	return cur, nil
}

func (s *segment) Read(off uint64) (*log_v1.Record, error) {
	_, pos, err := s.index.Read(int64(off - s.baseOffset))
	if err != nil {
		return nil, fmt.Errorf("error reading entry from index: %w", err)
	}
	p, err := s.store.Read(pos)
	if err != nil {
		return nil, fmt.Errorf("error reading record from store: %w", err)
	}
	record := &log_v1.Record{}
	err = proto.Unmarshal(p, record)
	return record, err
}

func (s *segment) IsMaxed() bool {
	return s.store.size >= s.config.Segment.MaxStoreBytes ||
		s.index.size >= s.config.Segment.MaxIndexBytes
}

func (s *segment) Remove() error {
	if err := s.Close(); err != nil {
		return err
	}
	if err := os.Remove(s.index.Name()); err != nil {
		return err
	}
	if err := os.Remove(s.store.Name()); err != nil {
		return err
	}
	return nil
}

func (s *segment) Close() error {
	if err := s.index.Close(); err != nil {
		return err
	}
	if err := s.store.Close(); err != nil {
		return err
	}
	return nil
}

func nearestMultiple(j, k uint64) uint64 {
	if j >= 0 {
		return (j / k) * k
	}
	return ((j - k + 1) / k) * k
}
